% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_theta_data.R
\name{clean_theta_data}
\alias{clean_theta_data}
\title{Clean ThetaProbe data for outliers in five minute data and save a CSV file of daily data}
\usage{
clean_theta_data(csv_in = NULL)
}
\arguments{
\item{csv_in}{A CSV file generated by \link{get_soil_moisture}, which should
be cleaned of outliers and aggregated to daily values}
}
\description{
This function checks for and removes outliers in the
University of Southern Queensland National Centre for Engineering in
Agriculture's (NCEA) Theta Probe data of soil moisture for Dr. Jo White,
saving the results as a CSV file with daily values
}
\details{
This function will check the CSV file for outliers in five minute
data, filter them and generate a new file with daily moisture values that will
be written to disk in the user's specified location.
}
\examples{
\dontrun{
cleaned <- clean_theta_data(csv = "~/Documents/Data/Theta Probe/2016-08-24_Soil_Moisture.csv",
path = "~/Documents/Data/Theta Probe/")

write.csv(cleaned, "~/Documents/cleaned_data.csv")
}
}

