% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_soilmoisture.R
\name{ftp_soilmoisture}
\alias{ftp_soilmoisture}
\title{Download, Clean and Generate Graphs From USQ Theta Probe Data}
\usage{
ftp_soilmoisture(userid = "", password = "")
}
\arguments{
\item{userid}{The login provided by NCEA to login via FTP}

\item{password}{The password provided by NCEA to login via FTP}
}
\description{
This package automates downloading and cleaning data from the
University of Southern Queensland National Centre for Engineering in
Agriculture (NCEA).

This is a slow process to retrieve all the files from the server.
}
\details{
This function will download all the .csv files from the server and
coallate them into one dataframe, save a .csv file to disk and visualise
the data.
}
\examples{

}

