% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Get Operating System Information}
\usage{
get_os()
}
\description{
This function checks the operating system being used and returns
a character string of the platform OS currently being used. Only for Windows
or MacOS/OSX.
}
\examples{

get_os()

}
