% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Clean ThetaProbe data for outliers in hourly data}
\usage{
remove_outliers(csv = NULL, path = NULL)
}
\arguments{
\item{csv}{A CSV file generated by get_soil_moisture()}

\item{path}{#' @param path Filepath to directory where a comma separated
file (CSV) exists, which must be generated by the get_soil_moisture()
function. This is the directory where the new CSV of daily values will be
placed as well.}
}
\description{
This function checks for and removes outliers in the
University of Southern Queensland National Centre for Engineering in
Agriculture's (NCEA) Theta Probe data of soil moisture for Dr. Jo White.
}
\details{
This function will check the CSV file for outliers in hourly data
and remove them and generate a new file with daily moisture values
}
\examples{
# NULL
}

