% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_moisture.R
\name{get_soil_moisture}
\alias{get_soil_moisture}
\title{Download, Clean and Generate Graphs From USQ Theta Probe Data}
\usage{
get_soil_moisture(userid = "", password = "", path = "")
}
\arguments{
\item{userid}{The login provided by NCEA to login via FTP}

\item{password}{The password provided by NCEA to login via FTP}

\item{path}{Filepath to directory for saving .csv file output. Defaults to
current working directory.}
}
\description{
This package automates downloading and cleaning data from the
University of Southern Queensland National Centre for Engineering in
Agriculture (NCEA).

This is a slow process to retrieve all the files from the server.
}
\details{
This function will download all the .csv files from the server and
coallate them into one dataframe named "soil_moisture" and saves a .csv file
to disk.
}
\examples{
# NULL
}

